requirements.txt
alembic
click
httptools
import-string
itsdangerous
Mako
manage.py
MarkupSafe
mimeparse
python-dateutil
sanic
six
SQLAlchemy
ujson
uvloop
websockets
jinja2
requests
aiohttp
psycopg2-binary
sanic_cors
ujson
aiohttp
aiosmtplib
redis
PyCrypto
bcrypt
git+https://github.com/gonrin/gatco.git
git+https://github.com/gonrin/gatco_auth.git
git+https://github.com/gonrin/gatco_sqlalchemy.git
git+https://github.com/gonrin/gatco_restapi.git
git+https://github.com/gonrin/gatco_acl.git
git+https://github.com/gonrin/gatco_jinja2.git
